+++ src/extensionVariables.ts
@@ -4,16 +4,19 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ExtensionContext, TreeView, StatusBarItem } from "vscode";
-import { KeyTar } from "./utils/keytar";
+// import { KeyTar } from "./utils/keytar";
+import * as keyTarType from "keytar";
 import { MemFS } from './fileSystemProvider'
 // import { } from './carTreeView'
+
+type KeyTar = typeof keyTarType;
 
 /**
  * Namespace for common variables used throughout the extension. They must be initialized in the activate() method of extension.ts
  */
 export namespace ext {
     export let context: ExtensionContext;
-    export let keytar: KeyTar | undefined;
+    export let keyTar: KeyTar;
     export let memFs: MemFS;
     export let hostStatusBar: StatusBarItem;
     export let carTreeData: object | undefined;