+++ src/utils/utils.ts
@@ -2,18 +2,17 @@ import * as vscode from 'vscode';
 import { ext } from '../extensionVariables';
 
 
-export function setHostStatusBar(host: string = '', password: string = '', tip: string = '') {
+export function setHostStatusBar(host: string = '', tip: string = '') {
 
     // // Create a status bar item
-	// const hostStatusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
-	// context.subscriptions.push(hostStatusBar);
+    // had ideas of passing in host details at connect, like tmos version and running ilx packages/versions
 
     ext.hostStatusBar.command = 'f5-fast.disconnect';
     ext.hostStatusBar.text = host ? host || '' : '';
-    ext.hostStatusBar.password = password ? password || '' : '';
+    // ext.hostStatusBar.password = password ? password || '' : '';
     ext.hostStatusBar.tooltip = tip ? tip || '' : '';
 
-    if (host && password) {
+    if (host) {
         ext.hostStatusBar.show();
     } else {
         ext.hostStatusBar.hide();
@@ -21,13 +20,10 @@ export function setHostStatusBar(host: string = '', password: string = '', tip: 
 
 };
 
-export function getHostStatusBar() {
-
-    // console.log(`hostStatusBar details: ${JSON.stringify(ext.hostStatusBar)}`);
-
-    return { host: ext.hostStatusBar.text, password: ext.hostStatusBar.password }
-
-};
+// export function getHostStatusBar() {
+//     // console.log(`hostStatusBar details: ${JSON.stringify(ext.hostStatusBar)}`);
+//     return { host: ext.hostStatusBar.text, password: ext.hostStatusBar.password }
+// };
 
 export function isValidJson(json: string) {
     try {
@@ -36,6 +32,28 @@ export function isValidJson(json: string) {
     } catch (e) {
         return false;
     }
+}
+
+export async function getPassword(device: string): Promise<any> {
+
+    // console.log(`getPassword Device: ${device}`);
+    
+    let password = await ext.keyTar.getPassword('f5Hosts', device).then( passwd => passwd )
+    
+    // console.log(`IS PASSWORD IN KEYTAR?: ${password}`);
+    if (!password) {
+        // console.log(`NO PASSWORD IN KEYTAR! - PROMPTING!!! - ${password}`);
+        password = await vscode.window.showInputBox({ password: true})
+        .then( password => {
+            if (!password) {
+                throw new Error('User cancelled password input')
+            }
+            // console.log(`USER INPUT PASSWORD!!! - ${password}`);
+            return password;
+            })
+    }
+    // console.log(`PASSWORD BOUT TO BE RETURNED!!! - ${password}`);
+    return password;
 }
 
 export function setMemento(key:string, value: string) {