+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 [BACK TO MAIN README](README.md)
 
-All notable changes to the "f5-fast" extension will be documented in this file.
+All notable changes to the "vscode-f5-fast" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+
+## [0.1.1] - (5-8-2020)
+
+### Added
+- password caching
 
 ## [0.1.0] - (5-7-2020)